Move factory methods to new builder class
