+++ app/src/main/java/com/amigo/user/UserBuilder.java
@@ -1,0 +1,71 @@
+package com.amigo.user;
+
+import com.amigo.course.Course;
+import com.amigo.form.CourseForm;
+import com.amigo.form.CourseFormList;
+import com.amigo.form.InterestForm;
+import com.amigo.form.RegistrationForm;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * Builder for {@code User} objects.
+ */
+@Service
+public class UserBuilder {
+    
+    private User user;
+    private Profile profile;
+    
+    public UserBuilder() {
+        this.reset();
+    }
+    
+    /**
+     * Populates the {@code User} object with attributes from the registration form.
+     */
+    public void populate(RegistrationForm form) {
+        profile.setName(form.getFirstName() + " " + form.getLastName());
+        user.setEmail(form.getEmail());
+        user.setPassword(form.getPassword());
+    }
+
+    /**
+     * Populates the {@code User} object with attributes from the course form.
+     */
+    public void populate(CourseFormList form) {
+        for (var each : form.getCourseList()) {
+            populateCourse(each);
+        }
+    }
+    
+    private void populateCourse(CourseForm form) {
+        Course c = new Course(form.getCourseCode(), form.getLectureCode(), form.getTutorialCode());
+        profile.addCourses(c);
+    }
+
+    /**
+     * Populates the {@code User} object with attributes from the interest form.
+     */
+    public void populate(InterestForm form) {
+        profile.setHobbies(form.getHobbies());
+        profile.setRecInterest(form.getRecreational());
+        profile.setMusInterest(form.getMusic());
+        profile.setSportInterest(form.getSports());
+    }
+    
+
+    public User createUser() {
+        user.setProfile(this.profile);
+        reset();
+        return user; 
+    }
+    
+    /**
+     * Discards all previous attributes.
+     */
+    private void reset() {
+        user = new User();
+        profile = new Profile();
+    }
+}
